requirements.txt
pyTelegramBotAPI~=4.13.0
python-dotenv~=0.21.1
telebot~=0.0.4
peewee~=3.16.2
telegraph~=2.2.0
boto~=2.49.0
boto3~=1.26.157
s3file~=0.0.0
Flask~=2.3.2
Pillow~=10.0.0
xlrd==2.0.1
YooMoney==0.1.0